+++ modules/seedvr/src/models/video_vae_v3/modules/attn_video_vae.py
@@ -1044,7 +1044,7 @@ class VideoAutoencoderKL(diffusers.AutoencoderKL):
         norm_num_groups: int = 32,
         sample_size: int = 32,
         scaling_factor: float = 0.18215,
-        force_upcast: float = True,
+        force_upcast: float = False,
         attention: bool = True,
         temporal_scale_num: int = 0,
         slicing_up_num: int = 0,
@@ -1244,12 +1244,14 @@ class VideoAutoencoderKL(diffusers.AutoencoderKL):
         blend_extent = int(self.tile_latent_min_size * self.tile_overlap_factor)
         row_limit = self.tile_latent_min_size - blend_extent
         rows = []
+        self.tiles = 0
         for i in range(0, x.shape[3], overlap_size):
             row = []
             for j in range(0, x.shape[4], overlap_size):
                 tile = x[:, :, :, i : i + self.tile_sample_min_size, j : j + self.tile_sample_min_size]
                 tile = self._encode(tile)
                 row.append(tile)
+                self.tiles += 1
             rows.append(row)
         result_rows = []
         for i, row in enumerate(rows):
@@ -1329,6 +1331,7 @@ class VideoAutoencoderKLWrapper(VideoAutoencoderKL):
         self.spatial_downsample_factor = spatial_downsample_factor
         self.temporal_downsample_factor = temporal_downsample_factor
         self.freeze_encoder = freeze_encoder
+        self.freeze_encoder = True
         super().__init__(*args, **kwargs)
 
     def forward(self, x: torch.FloatTensor) -> CausalAutoencoderOutput: